+++ examples/files.json
@@ -374,7 +374,7 @@
 		"webgpu_mirror",
 		"webgpu_multisampled_renderbuffers",
 		"webgpu_materials_texture_anisotropy",
-		"webgpu_storage_pbo"
+		"webgpu_storage_sorting"
 	],
 	"webaudio": [
 		"webaudio_orientation",
