+++ examples/jsm/nodes/core/NodeBuilder.js
@@ -1034,6 +1034,18 @@ class NodeBuilder {
 
 	}
 
+	setupPBONode( /*shaderStage*/ ) {
+
+		console.warn( 'Abstract function.' );
+
+	}
+
+	getPBOUniform( /*shaderStage*/ ) {
+
+		console.warn( 'Abstract function.' );
+
+	}
+
 	getCodes( shaderStage ) {
 
 		const codes = this.codes[ shaderStage ];
